Final_Project/requirements.txt
gradio_webrtc==0.0.28
google-genai==0.3.0
twilio
gradio
numpy
Pillow
python-dotenv
pyaudio
opencv-python
pyyaml